+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## \[Unreleased\]
+
 ### Added
 - Types `ExternalHash` and `AnyLinkableHash`.
+- New Admin API endpoint `storageInfo`.
 ### Changed
 ### Fixed
 ### Removed